portfolio_gain/pyproject.toml
[tool.poetry]
name = "portfolio_gain"
version = "0.1.0"
description = ""
authors = ["Dima <dima.gonikman@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
yfinance = "^0.1.63"
beautifulsoup4 = "^4.9.3"
selenium = "^3.141.0"
CurrencyConverter = "^0.16.4"
colorama = "^0.4.4"
Flask = "^2.0.1"
idna = "^3.2"
python-telegram-bot = "^13.7"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
